# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "API de gestão de utilizadores e testes de COVID-19",
        "description": "API que faz a gestão de utilizadores (técnicos, clientes e admins) e de testes de covid-19 de um centro de analises - Efetuar login para usar a API"
    },
    "host": "localhost:3000",
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Efetuar login",
                "parameters": [
                    {
                        "in": "body",
                        "name": "Credenciais",
                        "description": "Credenciais de Login",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "400": {
                        "description": "Login Failed"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Efetuar Logout",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Logout success"
                    },
                    "400": {
                        "description": "Logout Failed"
                    }
                }
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Obter utilizador que fez login",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Failed"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "id": {
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
        "Covtest": {
            "required": [
                "code"
            ],
            "properties": {
                "_id": {
                    "type": "string",
                    "uniqueItems": true
                },
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "userHistory": {
                    "type": "string"
                },
                "userStatus": {
                    "type": "string"
                },
                "riskGroup": {
                    "type": "string"
                },
                "riskJob": {
                    "type": "string"
                },
                "testStatus": {
                    "type": "string"
                },
                "testResult": {
                    "type": "string"
                },
                "resultFile": {
                    "type": "string"
                },
                "isTestDone": {
                    "type": "boolean"
                },
                "schedule": {
                    "type": "string",
                    "format": "date"
                },
                "user": {
                    "type": "object",
                    "$ref": "#/definitions/User"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date",
                    "default": "Date.now"
                }
            }
        },
        "User": {
            "required": [
                "email",
                "password",
                "idCard"
            ],
            "properties": {
                "_id": {
                    "type": "string",
                    "uniqueItems": true
                },
                "name": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "age": {
                    "type": "number"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "number"
                },
                "idCard": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "covtest": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Covtest"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "format": "date",
                    "default": "Date.now"
                }
            }
        },
...
    }
}